git-svn-id: svn://svn.r-forge.r-project.org/svnroot/unfoldr@106 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/unfoldr/src/Intersector.h
@@ -203,7 +203,7 @@ namespace STGM
       CEllipse3 & getEllipse () { return m_ellipse; };
       const CEllipse3 & getEllipse () const { return m_ellipse; };
 
-      int getType() const { return m_type; };
+      int getType() const { return (int) m_type; };
       int getSide() const { return m_side; };
 
       CPlane & getPlane() { return m_plane; }
